move vite alias resolver to folder
