+++ apps/desktop/vite.config.ts
@@ -2,7 +2,7 @@ import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
 import svgr from 'vite-plugin-svgr';
 import tsconfigPaths from 'vite-tsconfig-paths';
-import { relativeAliasResolver } from "@sd/config"
+import { relativeAliasResolver } from "@sd/config/vite"
 import { name, version } from './package.json';
 
 // https://vitejs.dev/config/
