+++ packages/config/package.json
@@ -1,9 +1,10 @@
 {
 	"name": "@sd/config",
 	"version": "0.0.0",
-	"main": "index.js",
-	"types": "index.d.ts",
 	"license": "GPL-3.0-only",
+	"exports": {
+		"./vite": "./vite/index.js"
+	},
 	"files": [
 		"eslint-react.js"
 	],
