--- packages/config/tsconfig.json
@@ -1,9 +1,0 @@
-{
-	"extends": "../config/base.tsconfig.json",
-	"include": [
-		"index.ts"
-	],
-	"compilerOptions": {
-		"outDir": "dist"
-	}
-}
