+++ packages/config/vite/index.d.ts
