+++ packages/config/vite/index.js
